Create LICENSE
This commit adds a LICENSE file to the repository, introducing the MIT License as our chosen framework for open-source collaboration. By adopting the MIT License, we clarify the terms under which our code may be used, modified, and distributed. This permissive license not only safeguards contributors and users by setting clear legal guidelines, but it also encourages an open and collaborative development environment.